feat: completed API implementation

+++ TestScoring.Domain/Entities/TestScoreFile.cs
@@ -3,11 +3,11 @@
 public class TestScoreFile : Entity
 {
     public string FileName { get; private set; }
-    public decimal FileSize { get; private set; }
+    public long FileSize { get; private set; }
     public string  FileExtenstion { get; private set; }
     public int Rows { get; private set; }
 
-    public TestScoreFile(string fileName, decimal fileSize, string fileExtenstion, int rows)
+    public TestScoreFile(string fileName, long fileSize, string fileExtenstion, int rows)
     {
         FileName = fileName;
         FileSize = fileSize;
